add namespase in controller
