+++ app/assets/stylesheets/custom/card.scss
@@ -1,0 +1,3 @@
+.card-weight {
+  width: 18rem;
+}
