+++ app/controllers/application_controller.rb
@@ -1,6 +1,8 @@
 class ApplicationController < ActionController::Base
   before_action :configure_permitted_parameters, if: :devise_controller?
 
+  add_flash_types :success, :danger, :warning
+
   protected
 
   def configure_permitted_parameters
