+++ app/views/application/_navigation.html.erb
@@ -10,12 +10,15 @@
         <li class="nav-item">
           <%= link_to "Home", root_path, class: "nav-link" %>
         </li>
+        <li class="nav-item">
+          <%= link_to "Posts", posts_path, class: "nav-link" %>
+        </li>
         <% if user_signed_in? %>
           <li class="nav-item">
-            <%= link_to "My Posts", user_posts_path(current_user), class: "nav-link" %>
+            <%= link_to current_user.email, user_path(current_user), class: "nav-link" %>
           </li>
           <li class="nav-item">
-            <%= link_to current_user.email, user_path(current_user), class: "nav-link" %>
+            <%= link_to "My Posts", user_users_posts_path(current_user), class: "nav-link" %>
           </li>
           <li class="nav-item">
             <%= link_to "Sign out", destroy_user_session_path, data: { turbo_method: :delete }, class: "nav-link" %>
