+++ app/views/layouts/application.html.erb
@@ -15,9 +15,13 @@
   </header>
 
   <body>
-    <p class="notice"><%= notice %></p>
-    <p class="alert"><%= alert %></p>
+    <% flash.each do |type, msg| %>
+      <div class="alert alert-<%= type %>">
+        <%= msg %>
+      </div>
+    <% end %>
 
-    <%= yield %>
+
+  <%= yield %>
   </body>
 </html>
