+++ app/views/posts/index.html.erb
@@ -1,8 +1,3 @@
-<p class="alert alert-success"><%= notice %></p>
-
-<h1>Posts</h1>
-<%= link_to "New post", new_user_post_path(@user) %>
-
 <div id="posts">
   <table class="table">
     <thead>
@@ -13,24 +8,12 @@
     </tr>
     </thead>
     <tbody>
-    <% if @user == current_user %>
-      <% @posts.each do |post| %>
-        <tr>
-          <th scope="row"><%= post.id %></th>
-          <td><%= link_to post.title, user_post_path(@user, post) %></td>
-          <td><%= post.description %></td>
-          <td><%= post.status %></td>
-        </tr>
-      <% end %>
-    <% else %>
-      <% @posts.each do |post| %>
-        <% next if post.status == "Draft" %>
-        <tr>
-          <th scope="row"><%= post.id %></th>
-          <td><%= link_to post.title, user_post_path(@user, post) %></td>
-          <td><%= post.description %></td>
-        </tr>
-      <% end %>
+    <% @posts.each do |post| %>
+      <tr>
+        <th scope="row"><%= post.id %></th>
+        <td><%= link_to post.title, user_users_post_path(post.user_id, post) %></td>
+        <td><%= post.description %></td>
+      </tr>
     <% end %>
     </tbody>
   </table>
