+++ app/views/users/index.html.erb
@@ -1,5 +1,3 @@
-<p class="alert alert-success"><%= notice %></p>
-
 <h1>Users</h1>
 
 <table class="table">
@@ -12,6 +10,3 @@
   <% end %>
   </tbody>
 </table>
-
-<div id="users">
-</div>
