+++ app/views/users/posts/edit.html.erb
@@ -5,5 +5,5 @@
 <br>
 
 <div>
-  <%= link_to "Back to posts", user_posts_path(@user) %>
+  <%= link_to "Back to posts", user_users_posts_path(@user) %>
 </div>
