+++ app/views/users/posts/index.html.erb
@@ -1,0 +1,26 @@
+<h1>Posts</h1>
+<% if current_user.present? %>
+    <%= link_to "New post", new_user_users_post_path(current_user) %>
+<% end %>
+
+<div id="posts">
+  <table class="table">
+    <thead>
+    <tr>
+      <th scope="col">#</th>
+      <th scope="col">Title</th>
+      <th scope="col">Description</th>
+    </tr>
+    </thead>
+    <tbody>
+      <% @posts.each do |post| %>
+        <tr>
+          <th scope="row"><%= post.id %></th>
+          <td><%= link_to post.title, user_users_post_path(post.user_id, post) %></td>
+          <td><%= post.description %></td>
+          <td><%= post.status %></td>
+        </tr>
+      <% end %>
+    </tbody>
+  </table>
+</div>
