+++ app/views/users/posts/show.html.erb
@@ -1,5 +1,3 @@
-<p class="alert alert-success"><%= notice %></p>
-
 <div class="card weight">
   <div class="card-body">
     <h5 class="card-title"><%= @post.title %></h5>
@@ -9,9 +7,9 @@
     <p class="card-text"><%= @post.status %></p>
 
     <div>
-      <%= link_to "Edit this post", edit_user_post_path(@user, @post) %> |
-      <%= link_to "Back to posts", user_posts_path(@user) %>
-      <%= button_to "Destroy this post", user_post_path(@user, @post), method: :delete %>
+      <%= link_to "Edit this post", edit_user_users_post_path(@user, @post) %> |
+      <%= link_to "Back to posts", user_users_posts_path(@user) %>
+      <%= button_to "Destroy this post", user_users_post_path(@user, @post), method: :delete %>
     </div>
   </div>
 </div>
