+++ app/views/users/show.html.erb
@@ -1,5 +1,3 @@
-<p class="alert alert-success"><%= notice %></p>
-
 <%= link_to "Back", root_path %>
 
 <% if @user == current_user %>
@@ -14,7 +12,7 @@
       <p class="card-text"><%= @user.surname %></p>
 
       <%= link_to "Refresh Credentials", edit_user_registration_path %>
-      <%= button_to "My Posts", "#" %>
+      <%= link_to "My Posts", user_users_posts_path(@user) %>
       <%= button_to "Deactivate Account", "#", method: :delete %>
     </div>
   </div>
@@ -23,7 +21,7 @@
 <div>
   <%= link_to "Back to users", users_path %>
 
-  <%= link_to "Posts", user_posts_path(@user) %>
+  <%= link_to "My Posts", user_users_posts_path(@user) %>
 
   <% if current_user == @user %>
     <%= link_to "Refresh Credentials", edit_user_registration_path %>
