+++ config/routes.rb
@@ -1,11 +1,12 @@
 Rails.application.routes.draw do
   root to: 'users#index'
   devise_for :users
-  resources :users, only: %i[index show destroy] do
-    resources :posts
-  end
-  # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
 
-  # Defines the root path route ("/")
-  # root "articles#index"
+  get '/posts', to: 'posts#index'
+
+  resources :users, only: %i[index show destroy] do
+    namespace :users do
+      resources :posts
+    end
+  end
 end
